Add ability to retrieve all nested values with a wildcard parent key
Today, `PropertyFiltering` supports filtering:
- at the top level of an object
- at the top level of objects in a list
- for nested objects, but only where the outer key is known in advance

This generally does’t work if the response type is a Map with keys that
are not known in advance. This is the typical pattern for the response
to a multiget - you return each object keyed by its ID. This PR
addresses that requirement.

There is a new wildcard (`*`) that can be passed to property filtering.
This now returns just the properties inside the nested objects, plus
the key to these objects.

So for a JSON:
```
{
   "bhalligan":{
      "firstname":"Brian",
      "lastname":"Halligan"
   },
   "spurcell":{
      "firstname":"Stephen",
      "lastname":"Purcell"
   }
}
```

And property filter: `*.firstname`, the result would be:
```
{
   "bhalligan":{
      "firstname":"Brian"
   },
   "spurcell":{
      "firstname":"Stephen"
   }
}
```